requirements.txt
streamlit==1.36.0
opencv-python==4.10.0.84
numpy>=2.0,<3.0
Pillow>=10.0.0
deepface==0.0.93
tensorflow>=2.20.0
matplotlib==3.9.2
pandas>=2.0.0